logo ve isim değiştirildi, fees kısmı API dan çekilerek test edildi, QR kod kısmına yarın bakılacak

+++ lib/Modules/Models/Fee.dart
@@ -26,9 +26,9 @@ class Fee{
       flatId: json['FLATID'] ?? 0,
       feeTypeId: json['FEETYPEID'] ?? 0,
       feeDate: json['FEEDATE'] ?? '',
-      feeAmount: json['FEEAMOUNT'] ?? 0.0,
+      feeAmount: (json['FEEAMOUNT'] ?? 0.0).toDouble(),
       paymentDate: json['PAYMENTDATE'] ?? '',
-      paymentAmount: json['PAYMENTAMOUNT'] ?? 0.0,
+      paymentAmount: (json['PAYMENTAMOUNT'] ?? 0.0).toDouble(),
     );
   }
 }
